refactor: rename registry.py to transforms.py
